Massive refactoring of components: Components renamed according to Vue style guide Much greater simplification and code reuse Greater use of slots Greater use of components Code cleanup and misc. bugfixes
