+++ client/src/App.vue
@@ -1,14 +1,41 @@
 <template>
-  <!-- <component
+  <div
+    v-if="owner.ownerId"
+    class="flex flex-col min-h-screen w-full px-2 py-2 sm:px-6 sm:py-4 xl:px-12 xl:py-6"
+  >
+    <!-- <component
     :is="view"
     :owner="owner"
     :images="images"
     :usage="usage"
   ></component> -->
+    <!-- <div v-if="owner.ownerId"> -->
+    <router-view :owner="owner" :images="images" :usage="usage"></router-view>
+    <!-- </div> -->
+    <!-- <div v-else class="fixed m-auto">
+    <svg
+      class="animate-spin mx-auto h-24 w-24 text-purple-500"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+    >
+      <circle
+        class="opacity-25"
+        cx="12"
+        cy="12"
+        r="10"
+        stroke="currentColor"
+        stroke-width="4"
+      ></circle>
+      <path
+        class="opacity-75"
+        fill="currentColor"
+        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+      ></path>
+    </svg>
+  </div> -->
 
-  <router-view :owner="owner" :images="images" :usage="usage"></router-view>
-
-  <!-- <router-view v-slot="{ Component }">
+    <!-- <router-view v-slot="{ Component }">
     <transition name="slide" mode="out-in">
       <keep-alive>
         <component
@@ -20,11 +47,12 @@
       </keep-alive>
     </transition>
   </router-view> -->
+  </div>
 </template>
 
 <script>
 import axios from 'axios';
-import { ref, shallowRef, reactive, provide } from 'vue';
+import { ref, reactive, provide } from 'vue';
 import { Notyf } from 'notyf';
 import 'notyf/notyf.min.css';
 import { useRoute } from 'vue-router';
@@ -35,29 +63,21 @@ import Account from './views/Account';
 export default {
   // components: { Home },
   setup() {
-    console.log('in App');
     const route = useRoute();
     const server = process.env.VUE_APP_SERVER;
     const isReadyToRender = ref(false);
-    // const view = shallowRef(Home);
-
-    // provide('changeView', newView => {
-    //   console.log('view.value:', view.value);
-    //   view.value = newView;
-    //   console.log('view.value:', view.value);
-    // });
 
     const owner = reactive({});
     const usage = ref('');
     const images = ref([]);
 
     provide('owner', owner);
-    provide('usage', usage.value);
     provide('images', images.value);
 
     // -------------------------------------
     // ------ INITIAL CONFIGURATION --------
     // ------------ & COOKIES --------------
+    // -------------------------------------
 
     const ownerId = getCookie('ownerId');
     const guestId = getCookie('guestId');
@@ -115,11 +135,11 @@ export default {
       switch (userType) {
         case 'owner':
           url = `${server}/auth/check-session`;
-          id = { ownerId: ownerId };
+          id = { ownerId };
           break;
         case 'guest':
           url = `${server}/user/get-owner`;
-          id = { guestId: guestId };
+          id = { guestId };
           break;
       }
 
@@ -137,8 +157,8 @@ export default {
         method: 'post',
         data: id
       });
-
-      if (!response.data.isLoggedIn) {
+      
+      if (userType === 'owner' && !response.data.isLoggedIn) {
         window.location = `${server}/login`;
       }
       images.value = response.data.images;
@@ -167,6 +187,7 @@ export default {
     // }
     // provide('changeHomeView', changeHomeView);
 
+    // Define default toast configuration and provide
     const toast = new Notyf({
       position: { x: 'right', y: 'top' },
       duration: 2000,
