+++ client/src/components/BaseButtonCancel.vue
@@ -1,0 +1,7 @@
+<template>
+  <button
+    class="w-36 px-3 py-1 bg-gray-600 rounded shadow text-center text-white hover:shadow-lg hover:bg-gray-500 transition duration-200 ease-in-out"
+  >
+    <slot></slot>
+  </button>
+</template>
