+++ client/src/components/BaseSkeletonImage.vue
@@ -1,0 +1,3 @@
+<template>
+  <div class="w-64 h-48 m-1 bg-gray-300 rounded-sm animate-pulse"></div>
+</template>
