+++ client/src/components/HomeGalleryEmpty.vue
@@ -1,0 +1,12 @@
+<template>
+  <div
+    
+    class="flex flex-grow justify-center rounded border-4 border-dashed border-gray-400 text-gray-500 hover:bg-gray-200 hover:text-gray-600 cursor-pointer transition-all duration-200 ease-in-out"
+  >
+    <p class="flex-grow self-center text-center text-2xl">
+      Upload some images!
+    </p>
+  </div>
+</template>
+
+
