+++ client/src/router/index.js
@@ -3,8 +3,8 @@ import { createRouter, createWebHistory } from 'vue-router';
 const Home = () => import('../views/Home.vue');
 // import Account from '../views/Account.vue';
 const Account = () => import('../views/Account');
-const AccountSummary = () => import('../views/AccountSummary');
-const ChangePlan = () => import('../views/ChangePlan');
+// const AccountSummary = () => import('../views/AccountSummary');
+// const AccountPlanPicker = () => import('../views/AccountPlanPicker');
 
 const routes = [
   {
@@ -14,22 +14,22 @@ const routes = [
     props: true
   },
   {
-    path: '/account/:ownerId',
+    path: '/account',
     name: 'account',
     component: Account,
-    props: true,
-    children: [
-      // {
-      //   path: '/',
-      //   component: AccountSummary,
-      //   props: true
-      // },
-      {
-        path: 'change-plan',
-        component: ChangePlan,
-        props: true
-      }
-    ]
+    props: true
+    // children: [
+    //   // {
+    //   //   path: '/',
+    //   //   component: AccountSummary,
+    //   //   props: true
+    //   // },
+    //   {
+    //     path: 'account-plan-picker',
+    //     component: AccountPlanPicker,
+    //     props: true
+    //   }
+    // ]
   }
 ];
