+++ client/src/views/Account.vue
@@ -1,325 +1,5 @@
 <template>
-  <!-- <div
-    v-if="showAccountSummary"
-    ref="parent"
-    class="flex flex-col min-h-screen w-full px-2 py-2 sm:px-6 sm:py-4 xl:px-12 xl:py-6"
-  >
-    <div class="flex flex-col sm:flex-row sm:justify-between">
-      <a href="/">
-        <colorful-logo
-          class="self-center sm:self-start"
-          :content="'carousel'"
-        />
-        <h1
-          class="text-lg tracking-tight leading-10 font-semibold text-gray-900 sm:text-lg sm:leading-none md:text-xl"
-        >
-          <span class="inline-block align-baseline text-gray-700 text">
-            Account: {{ owner.firstName }} {{ owner.lastName }}
-          </span>
-        </h1>
-      </a>
-      <div
-        class="flex flex-col items-center sm:items-end py-4 gap-y-4 mt-3 text-gray-700"
-      >
-        <router-link
-          to="/"
-          class="flex items-center space-x-2 border border-transparent transition-colors duration-150 focus:outline-none self-center px-5 py-1 font-medium leading-5 text-white bg-purple-500 rounded-lg shadow active:bg-purple-600 hover:bg-purple-600 focus:shadow-outline-purple cursor-pointer"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            class="w-4 h-4"
-            viewBox="0 0 20 20"
-            fill="currentColor"
-          >
-            <path
-              fill-rule="evenodd"
-              d="M9.707 16.707a1 1 0 01-1.414 0l-6-6a1 1 0 010-1.414l6-6a1 1 0 011.414 1.414L5.414 9H17a1 1 0 110 2H5.414l4.293 4.293a1 1 0 010 1.414z"
-              clip-rule="evenodd"
-            />
-          </svg>
-          <span>Back to Gallery</span>
-        </router-link>
-      </div>
-    </div>
-    <div
-      id="subscription-cancelled"
-      class="flex flex-wrap font-bold text-pasha text-xl mt-6 mb-2 hidden"
-      :class="elementClasses.subscriptionCancelled"
-    >
-      Subscription canceled
-    </div>
-    <div>
-      <button
-        class="bg-pasha hover:bg-white hover:shadow-outline hover:text-pasha hover:border hover:border-black focus:shadow-outline text-white focus:bg-white focus:text-pasha font-light py-2 px-4 rounded"
-        type="button"
-        onclick="window.location.href='/'"
-      >
-        Restart demo
-      </button>
-    </div>
-    <div
-      id="subscription-settings"
-      :class="elementClasses.subscriptionSettings"
-    >
-      <div class="flex flex-wrap justify-center mt-4">
-        <div class="md:w-2/5 w-full inline-block rounded-md p-4">
-          <div
-            id="subscription-status-text"
-            class="text-center font-bold text-pasha text-2xl"
-          >
-            Account settings
-          </div>
-          <div class="mt-4 border rounded p-4">
-            <div class="font-bold text-xl mb-2">
-              Account
-            </div>
-            <div class="flex justify-between text-gray-600 text-xl">
-              <div>Current plan</div>
-              <div id="subscribed-price" class="mb-2 font-bold text-xl">
-                {{ planDetails.plan }}
-              </div>
-            </div>
-
-            <div class="flex justify-between">
-              <div class="text-xl text-gray-600">
-                Credit card
-              </div>
-              <span
-                id="credit-card-last-four"
-                class="font-bold text-xl text-gray-600"
-              >
-                {{ planDetails.paymentMethod }}
-              </span>
-            </div>
-
-            <div
-              class="flex justify-between mt-2 mb-2 text-gray-900 font-bold text-xl cursor-pointer"
-              @click="updateBilling"
-            >
-              <span>
-                Update billing info
-                <span>→</span>
-              </span>
-            </div>
-            <div
-              class="flex justify-between mt-2 mb-2 text-gray-900 font-bold text-xl cursor-pointer"
-              @click="openPlanChange"
-            >
-              <span>
-                Change plan
-                <span>→</span>
-              </span>
-            </div>
-            <div
-              class="flex justify-between mt-2 mb-2 text-gray-900 font-bold text-xl cursor-pointer"
-              @click="cancelSubscription"
-            >
-              <span>
-                Cancel subscription
-                <span>→</span>
-              </span>
-            </div>
-          </div>
-
-          <div
-            id="prices-form"
-            class="w-full mr-4 md:mb-8"
-            :class="elementClasses.pricesForm"
-          >
-            <div class="text-center text-pasha font-bold text-2xl mt-4 mb-6">
-              Change pricing plan
-            </div>
-
-            <div class="flex justify-between mb-8">
-              <div
-                id="basic"
-                class="w-1/2 rounded overflow-hidden border p-2 mr-4 md:mr-8"
-              >
-                <div class="px-2 py-2">
-                  <div class="text-gray-500 text-xl mb-2 font-medium">
-                    Basic
-                  </div>
-                  <p class="text-pasha text-2xl font-extrabold">
-                    $5.00
-                  </p>
-                  <div class="flex-wrap">
-                    <div class="leading-none text-gray-500 text-xs font-medium">
-                      Per month
-                    </div>
-                    <div
-                      class="leading-none text-gray-500 text-xs font-medium mt-1"
-                    >
-                      Billed monthly
-                    </div>
-                  </div>
-
-                  <div class="flex justify-center mt-6">
-                    <button
-                      id="submit-basic"
-                      onClick="switchPrices('BASIC')"
-                      class="bg-pasha hover:bg-white outline-none hover:text-pasha hover:border hover:border-black text-white focus:bg-white focus:text-pasha font-light py-2 px-4 rounded-lg"
-                      type="submit"
-                    >
-                      <div
-                        id="submit-basic-button-text"
-                        class="w-auto -mx-2 md:mx-0"
-                      >
-                        <div id="loading" class=""></div>
-                        <span id="button-text" :class="elementClasses.btnText">
-                          Select
-                        </span>
-                      </div>
-                    </button>
-                  </div>
-                </div>
-              </div>
-              <div
-                id="premium"
-                class="w-1/2 rounded overflow-hidden border p-2"
-              >
-                <div class="px-2 py-2">
-                  <div class="text-gray-500 text-xl mb-2 font-medium">
-                    Premium
-                  </div>
-                  <p class="text-pasha text-2xl font-extrabold">
-                    $15.00
-                  </p>
-                  <div class="flex-wrap">
-                    <div class="leading-none text-gray-500 text-xs font-medium">
-                      Per month
-                    </div>
-                    <div
-                      class="leading-none text-gray-500 text-xs font-medium mt-1"
-                    >
-                      Billed monthly
-                    </div>
-                  </div>
-                  <div class="flex justify-center mt-6">
-                    <button
-                      id="submit-premium"
-                      onClick="switchPrices('PREMIUM')"
-                      class="bg-pasha hover:bg-white outline-none hover:text-pasha hover:border hover:border-black text-white focus:bg-white focus:text-pasha font-light py-2 px-4 rounded-lg"
-                      type="submit"
-                    >
-                      <div
-                        id="submit-premium-button-text"
-                        class="w-auto -mx-2 md:mx-0"
-                      >
-                        <div id="loading" class=""></div>
-                        <span id="button-text">Select</span>
-                      </div>
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div id="price-change-form" class="flex justify-center">
-              <div class="w-full rounded overflow-hidden border p-4 mb-4">
-                <div class="flex justify-between text-gray-600 text-m">
-                  <div>Current price</div>
-                  <div
-                    id="current-price-subscribed"
-                    class="font-bold text-m mb-2"
-                  ></div>
-                </div>
-
-                <div class="flex justify-between text-gray-600 text-m">
-                  <div>New price</div>
-                  <div
-                    id="new-price-selected"
-                    class="mb-2 font-bold text-m"
-                  ></div>
-                </div>
-
-                <div>
-                  <p class="mt-4 mb-4 text-gray-600">
-                    You will be charged
-                    <span id="new-price-price-selected"></span>
-                    on
-                    <span id="new-price-start-date"></span>
-                  </p>
-                  <button
-                    id="confirm-price-change-submit"
-                    onClick="confirmPriceChange()"
-                    class="bg-pasha hover:bg-white hover:shadow-outline hover:text-pasha hover:border hover:border-black focus:shadow-outline text-white focus:bg-white focus:text-pasha font-light py-2 px-4 rounded-lg"
-                    :class="elementClasses.confirmPriceChangeSubmit"
-                    type="submit"
-                  >
-                    <div class="w-auto -mx-2 md:mx-0">
-                      <div id="loading" class=""></div>
-                      <span id="button-text">Confirm change</span>
-                    </div>
-                  </button>
-                  <button
-                    id="confirm-price-change-cancel"
-                    onClick="cancelChangePrice()"
-                    class="bg-pasha hover:bg-white hover:shadow-outline hover:text-pasha hover:border hover:border-black focus:shadow-outline text-white focus:bg-white focus:text-pasha font-light py-2 px-4 rounded-lg"
-                    :class="elementClasses.confirmPriceChangeCancel"
-                    type="submit"
-                  >
-                    <div class="w-auto -mx-2 md:mx-0">
-                      <div id="loading" class=""></div>
-                      <span id="button-text">Cancel</span>
-                    </div>
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div> -->
-
-  <!-- <transition appear name="slide" mode="out-in">
-    <router-view
-      v-slot="{ Component }"
-      :current-plan="planDetails.plan"
-      @confirm-plan-change="confirmPlanChange"
-      @close-plan-change-modal="closePlanChangeModal"
-    >
-      <component :is="Component" />
-    </router-view>
-  </transition> -->
-  <div class="flex flex-col w-full px-2 pt-2 sm:px-6 sm:pt-4 xl:px-12 xl:pt-6">
-    <div class="flex flex-col sm:flex-row sm:justify-between">
-      <a href="/">
-        <colorful-logo
-          class="self-center sm:self-start"
-          :content="'carousel'"
-        />
-        <h1
-          class="text-lg tracking-tight leading-10 font-semibold text-gray-900 sm:text-lg sm:leading-none md:text-xl"
-        >
-          <span class="inline-block align-baseline text-gray-700 text">
-            Account: {{ owner.firstName }} {{ owner.lastName }}
-          </span>
-        </h1>
-      </a>
-      <div
-        class="flex flex-col items-center sm:items-end py-4 gap-y-4 mt-3 text-gray-700"
-      >
-        <router-link
-          to="/"
-          class="flex items-center space-x-2 border border-transparent transition-colors duration-150 focus:outline-none self-center px-5 py-1 font-medium leading-5 text-white bg-purple-500 rounded-lg shadow active:bg-purple-600 hover:bg-purple-600 focus:shadow-outline-purple cursor-pointer"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            class="w-4 h-4"
-            viewBox="0 0 20 20"
-            fill="currentColor"
-          >
-            <path
-              fill-rule="evenodd"
-              d="M9.707 16.707a1 1 0 01-1.414 0l-6-6a1 1 0 010-1.414l6-6a1 1 0 011.414 1.414L5.414 9H17a1 1 0 110 2H5.414l4.293 4.293a1 1 0 010 1.414z"
-              clip-rule="evenodd"
-            />
-          </svg>
-          <span>Back to Gallery</span>
-        </router-link>
-      </div>
-    </div>
-  </div>
+  <account-menu :owner="owner"></account-menu>
 
   <transition appear name="slide" mode="out-in">
     <component
@@ -334,9 +14,11 @@
 
 <script>
 import axios from 'axios';
-import ColorfulLogo from '../components/ColorfulLogo';
-import AccountSummary from './AccountSummary';
-import ChangePlan from './ChangePlan';
+import { defineAsyncComponent } from 'vue';
+import AccountMenu from '../components/AccountMenu';
+import AccountSummary from '../components/AccountSummary';
+import AccountPlanPicker from '../components/AccountPlanPicker';
+
 import {
   ref,
   shallowRef,
@@ -351,18 +33,13 @@ import { useRoute, useRouter } from 'vue-router';
 
 export default {
   name: 'Account',
-  components: { ColorfulLogo, AccountSummary, ChangePlan },
+  components: { AccountMenu, AccountSummary, AccountPlanPicker },
+  inheritAttrs: false,
   props: {
     owner: {
       type: Object,
       default: () => {
         return {};
-      }
-    },
-    images: {
-      type: Array,
-      default: () => {
-        return [];
       }
     }
   },
@@ -383,11 +60,6 @@ export default {
     elementClasses.pricesForm = 'hidden opacity-0';
 
     getCurrentPlan(owner.ownerId);
-    // onBeforeMount(async () => {
-    //   // planDetails.value = await getCurrentPlan(route.params.ownerId);
-    //   getCurrentPlan(route.params.ownerId);
-    //   // changeModalBtns(planDetails.value.plan);
-    // });
 
     async function getCurrentPlan() {
       const response = await axios.post(
@@ -423,9 +95,9 @@ export default {
     const showAccountSummary = ref(true);
 
     function openPlanChange() {
-      accountView.value = ChangePlan;
+      accountView.value = AccountPlanPicker;
 
-      // router.push('change-plan');
+      // router.push('account-plan-picker');
 
       // setTimeout(() => {
       //   showAccountSummary.value = false;
@@ -450,16 +122,48 @@ export default {
           newPriceId: newPriceId
         }
       );
-      getCurrentPlan();
-      closePlanChange();
+      console.log('response.data:', response.data);
+      const data = response.data;
+      if (
+        !data.subUpdated &&
+        data.msg.toLowerCase().includes('no such subscription')
+      ) {
+        toast.open({
+          type: 'info',
+          duration: 0,
+          dismissible: true,
+          message: '<div id="toast-message"><p id="msg-text"></p></div>'
+        });
+        document.getElementById('toast-message').innerText =
+          'Your subscription was not found. Please contact support or try again.';
+      } else if (!data.subUpdated) {
+        toast.open({
+          type: 'info',
+          duration: 0,
+          dismissible: true,
+          message: '<div id="toast-message"><p id="msg-text"></p></div>'
+        });
+        document.getElementById('msg-text').innerText =
+          'Your subscription could not be changed right now.\nPlease contact support or try again.';
+      }
+      if (data.subUpdated) {
+        getCurrentPlan(owner.ownerId);
+        closePlanChangeModal();
+
+        toast.open({
+          type: 'info',
+          duration: 5000,
+          // dismissible: true,
+          message: '<div id="toast-message"><p id="msg-text"></p></div>'
+        });
+        document.getElementById('msg-text').innerText =
+          'Subscription updated\n' + response;
+        // subscriptionCancelled(response);
+      }
       return response.data;
     }
 
     async function cancelSubscription() {
-      // changeLoadingStatePrices(true);
-      const params = new URLSearchParams(document.location.search.substring(1));
-      const subscriptionId = params.get('subscriptionId');
-
       const response = await axios(
         process.env.VUE_APP_SERVER + '/payment/cancel-subscription',
         {
